matched func_8001FAE0 25

+++ include/main/unknown_syms.h
@@ -56,7 +56,7 @@ extern void  func_8004E200(void);
 extern u8*   Midi_Event3(s32 arg0, u8* arg1, MidiSong* arg2, MidiTrack* arg3);
 extern s32   CdAudio_DriveNull(void);
 extern void  func_8001F430(void);
-extern s32   func_8001FAE0(s32 arg0, s32 arg1);
+extern s32   func_8001FAE0(u16 arg0, s32 arg1);
 extern void  func_80059EE0(void);
 extern s32   func_800AF590(s32 arg0, s32 arg1);
 extern s32   func_800B0118(s32 arg0, s32 arg1);
